+++ examples/typescript/http-server-pool/fastify/src/types.ts
@@ -1,0 +1,19 @@
+import { type ThreadPoolOptions } from 'poolifier'
+
+export interface BodyPayload {
+  number?: number
+}
+
+export interface WorkerData<T = unknown> {
+  body: T
+}
+
+export interface WorkerResponse<T = unknown> {
+  body: T
+}
+
+export interface FastifyPoolifierOptions extends ThreadPoolOptions {
+  workerFile: string
+  maxWorkers: number
+  minWorkers: number
+}